+++ main.go
@@ -18,28 +18,32 @@
 package main
 
 import (
+	"errors"
 	"fmt"
+	"net/http"
 	"os"
-	"runtime"
-	"time"
 
-	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/go-kit/kit/log"
+	"github.com/gorilla/mux"
+
 	"github.com/xmidt-org/argus/auth"
-	"github.com/xmidt-org/argus/store/db"
-	"github.com/xmidt-org/argus/store/db/metric"
-	"github.com/xmidt-org/themis/xmetrics/xmetricshttp"
+	"github.com/xmidt-org/arrange"
+	"github.com/xmidt-org/arrange/arrangehttp"
+	"github.com/xmidt-org/httpaux"
+	"github.com/xmidt-org/sallust/sallustkit"
+	"github.com/xmidt-org/themis/config"
+	"github.com/xmidt-org/themis/xmetrics"
+	"github.com/xmidt-org/touchstone"
 	"github.com/xmidt-org/webpa-common/basculechecks"
 	"github.com/xmidt-org/webpa-common/basculemetrics"
 
-	"github.com/InVisionApp/go-health"
+	"github.com/xmidt-org/touchstone/touchhttp"
+
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
-	"github.com/xmidt-org/themis/config"
-	"github.com/xmidt-org/themis/xhealth"
-	"github.com/xmidt-org/themis/xhttp/xhttpserver"
-	"github.com/xmidt-org/themis/xlog"
-	"github.com/xmidt-org/themis/xlog/xloghttp"
+
 	"go.uber.org/fx"
+	"go.uber.org/zap"
 )
 
 const (
@@ -53,103 +57,106 @@ var (
 	BuildTime = "undefined"
 )
 
-func setupFlagSet(fs *pflag.FlagSet) error {
-	fs.StringP("file", "f", "", "the configuration file to use.  Overrides the search path.")
-	fs.BoolP("debug", "d", false, "enables debug logging.  Overrides configuration.")
-	fs.BoolP("version", "v", false, "print version and exit")
-
-	return nil
-}
-
-func setupViper(in config.ViperIn, v *viper.Viper) (err error) {
-	if printVersion, _ := in.FlagSet.GetBool("version"); printVersion {
-		printVersionInfo()
-	}
-	if file, _ := in.FlagSet.GetString("file"); len(file) > 0 {
-		v.SetConfigFile(file)
-		err = v.ReadInConfig()
-	} else {
-		v.SetConfigName(string(in.Name))
-		v.AddConfigPath(fmt.Sprintf("/etc/%s", in.Name))
-		v.AddConfigPath(fmt.Sprintf("$HOME/.%s", in.Name))
-		v.AddConfigPath(".")
-		err = v.ReadInConfig()
-	}
-
-	if err != nil {
-		return
-	}
-
-	if debug, _ := in.FlagSet.GetBool("debug"); debug {
-		v.Set("log.level", "DEBUG")
-	}
-
-	return nil
-}
-
-func printVersionInfo() {
-	fmt.Fprintf(os.Stdout, "%s:\n", applicationName)
-	fmt.Fprintf(os.Stdout, "  version: \t%s\n", Version)
-	fmt.Fprintf(os.Stdout, "  go version: \t%s\n", runtime.Version())
-	fmt.Fprintf(os.Stdout, "  built time: \t%s\n", BuildTime)
-	fmt.Fprintf(os.Stdout, "  git commit: \t%s\n", GitCommit)
-	fmt.Fprintf(os.Stdout, "  os/arch: \t%s/%s\n", runtime.GOOS, runtime.GOARCH)
-	os.Exit(0)
-}
-
 func main() {
+	v, logger, err := setup(os.Args[1:])
+	if err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
 	app := fx.New(
-		xlog.Logger(),
-		config.CommandLine{Name: applicationName}.Provide(setupFlagSet),
-		provideMetrics(),
-		metric.ProvideMetrics(),
-		basculechecks.ProvideMetricsVec(),
+		arrange.LoggerFunc(logger.Sugar().Infof),
+		arrange.ForViper(v),
+		fx.Supply(logger),
+		fx.Supply(v),
+		touchstone.Provide(),
+		touchhttp.Provide(),
 		basculemetrics.ProvideMetricsVec(),
+		basculechecks.ProvideMetricsVec(),
 		auth.ProvidePrimaryServerChain(apiBase),
 		fx.Provide(
+			backwardsCompatibleMetricFactory("prometheus"),
+			backwardsCompatibleUnmarshaller,
+			backwardsCompatibleLogger,
 			auth.ProfilesUnmarshaler{
 				ConfigKey:        "authx.inbound.profiles",
 				SupportedServers: []string{"primary"}}.Annotated(),
-			config.ProvideViper(setupViper),
-			xlog.Unmarshal("log"),
-			xloghttp.ProvideStandardBuilders,
-			db.Provide,
-			xhealth.Unmarshal("health"),
-			provideServerChainFactory,
-			xmetricshttp.Unmarshal("prometheus", promhttp.HandlerOpts{}),
-			xhttpserver.Unmarshal{Key: "servers.primary"}.Annotated(),
-			xhttpserver.Unmarshal{Key: "servers.metrics"}.Annotated(),
-			xhttpserver.Unmarshal{Key: "servers.health"}.Annotated(),
+			arrange.UnmarshalKey("prometheus", touchstone.Config{}),
+			arrange.UnmarshalKey("prometheus.handler", touchhttp.Config{}),
+			metricMiddleware,
 		),
+
+		arrangehttp.Server{
+			Name: "server_primary",
+			Key:  "servers.primary",
+			Inject: arrange.Inject{
+				PrimaryMMIn{},
+			},
+		}.Provide(),
+
+		arrangehttp.Server{
+			Name: "server_health",
+			Key:  "servers.health",
+			Inject: arrange.Inject{
+				HealthMMIn{},
+			},
+			Invoke: arrange.Invoke{
+				func(r *mux.Router) {
+					r.Handle("/health", httpaux.ConstantHandler{
+						StatusCode: http.StatusOK,
+					}).Methods("GET")
+				},
+			},
+		}.Provide(),
+
+		arrangehttp.Server{
+			Name: "server_metrics",
+			Key:  "servers.metrics",
+		}.Provide(),
 
 		fx.Invoke(
 			serverValidator{Key: "servers.primary"}.Validate,
 			serverValidator{Key: "servers.metrics"}.Validate,
 			serverValidator{Key: "servers.health"}.Validate,
-			xhealth.ApplyChecks(
-				&health.Config{
-					Name:     applicationName,
-					Interval: 24 * time.Hour,
-					Checker: xhealth.NopCheckable{
-						Details: map[string]interface{}{
-							"StartTime": time.Now().UTC().Format(time.RFC3339),
-						},
-					},
-				},
-			),
-			BuildPrimaryRoutes,
-			BuildMetricsRoutes,
-			BuildHealthRoutes,
+			buildPrimaryRoutes,
+			buildMetricsRoutes,
 		),
 	)
 
-	switch err := app.Err(); err {
-	case pflag.ErrHelp:
+	switch err := app.Err(); {
+	case errors.Is(err, pflag.ErrHelp):
 		return
-	case nil:
+	case err == nil:
 		app.Run()
 	default:
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(2)
+	}
+}
+
+func backwardsCompatibleLogger(l *zap.Logger) log.Logger {
+	return sallustkit.Logger{
+		Zap: l,
+	}
+}
+
+func backwardsCompatibleUnmarshaller(v *viper.Viper) config.Unmarshaller {
+	return config.ViperUnmarshaller{
+		Viper: v,
+	}
+}
+
+func backwardsCompatibleMetricFactory(configKey string) func(xmetrics.MetricsIn) (xmetrics.Factory, error) {
+	return func(in xmetrics.MetricsIn) (xmetrics.Factory, error) {
+		var o xmetrics.Options
+		if err := in.Unmarshaller.UnmarshalKey(configKey, &o); err != nil {
+			return nil, err
+		}
+
+		registry, err := xmetrics.New(o)
+		if err != nil {
+			return nil, err
+		}
+
+		return registry, nil
 	}
 }