+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Start migrating away from Themis sub-packages. [#6](https://github.com/xmidt-org/authbaton/pull/6)
 
 ## [v0.0.1]
 - Initial creation